move outbound control number generation to partners
